adding autoBugTracker and executeUserScript.py
